+++ BE/exceed/src/main/java/com/gaebaljip/exceed/achieve/application/port/in/GetAchieveUsecase.java
@@ -1,6 +1,6 @@
 package com.gaebaljip.exceed.achieve.application.port.in;
 
-import com.gaebaljip.exceed.dto.GetAchieveListResponse;
+import com.gaebaljip.exceed.dto.response.GetAchieveListResponse;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
