+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/request/EatMealRequest.java
@@ -1,4 +1,4 @@
-package com.gaebaljip.exceed.dto;
+package com.gaebaljip.exceed.dto.request;
 
 import com.gaebaljip.exceed.common.annotation.Enum;
 import com.gaebaljip.exceed.meal.domain.MealType;
