+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/response/GetAchieveListResponse.java
@@ -1,4 +1,4 @@
-package com.gaebaljip.exceed.dto;
+package com.gaebaljip.exceed.dto.response;
 
 import lombok.Builder;
