+++ BE/exceed/src/main/java/com/gaebaljip/exceed/dto/response/GetFood.java
@@ -1,4 +1,4 @@
-package com.gaebaljip.exceed.dto;
+package com.gaebaljip.exceed.dto.response;
 
 public record GetFood(
         String name) {
