+++ BE/exceed/src/main/java/com/gaebaljip/exceed/meal/application/port/in/GetFoodQuery.java
@@ -1,6 +1,6 @@
 package com.gaebaljip.exceed.meal.application.port.in;
 
-import com.gaebaljip.exceed.dto.GetFood;
+import com.gaebaljip.exceed.dto.response.GetFood;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
