+++ BE/exceed/src/main/java/com/gaebaljip/exceed/meal/application/port/in/GetSpecificMealQuery.java
@@ -1,6 +1,6 @@
 package com.gaebaljip.exceed.meal.application.port.in;
 
-import com.gaebaljip.exceed.dto.GetMeal;
+import com.gaebaljip.exceed.dto.response.GetMeal;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
