+++ BE/exceed/src/main/java/com/gaebaljip/exceed/member/application/port/in/GetMaintainMealUsecase.java
@@ -1,6 +1,6 @@
 package com.gaebaljip.exceed.member.application.port.in;
 
-import com.gaebaljip.exceed.dto.MaintainMeal;
+import com.gaebaljip.exceed.dto.response.MaintainMeal;
 import org.springframework.stereotype.Component;
 
 @Component
