+++ BE/exceed/src/main/java/com/gaebaljip/exceed/member/application/port/in/GetTargetMealUsecase.java
@@ -1,6 +1,6 @@
 package com.gaebaljip.exceed.member.application.port.in;
 
-import com.gaebaljip.exceed.dto.TargetMeal;
+import com.gaebaljip.exceed.dto.response.TargetMeal;
 import org.springframework.stereotype.Component;
 
 @Component
